prettier classes command
`classes` command is now pretty and supports plain string filtering; the
AllClasses command has been replaced with AllClassesWithGeneric, which
returns generic classes and works with the DalvikVM; added the JavaClass
object for storage of the above.

added command callback type that supports a dictionary of arguments,
this is successfully used in the `classes` command.

fixed bug with exiting, wherein `onecmd` did not return what the command
really returned